% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omie2fdata-package.R
\docType{package}
\name{omie2fdata-package}
\alias{omie2fdata-package}
\title{omie2fdata: OMIE electricity-market data as functional data}
\description{
Tools to download, read and normalize public OMIE electricity-market files
and convert daily trajectories into \code{fdata} and \code{ldata} objects from
the \code{fda.usc} package.
}
\details{
The package separates raw acquisition from conversion:
\itemize{
\item \code{\link[=omie2txt]{omie2txt()}} downloads the consolidated daily market-results table.
\item \code{\link[=omie2download]{omie2download()}} downloads daily files from OMIE's public file repository.
\item \code{\link[=omie2df]{omie2df()}} provides a common tabular layer.
\item \code{\link[=omie2fdata]{omie2fdata()}} and \code{\link[=omie2ldata]{omie2ldata()}} create functional-data objects.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://moviedo5.github.io/omie2fdata}
  \item \url{https://github.com/moviedo5/omie2fdata}
  \item Report bugs at \url{https://github.com/moviedo5/omie2fdata/issues}
}

}
\author{
\strong{Maintainer}: Manuel Oviedo de la Fuente \email{manuel.oviedo@udc.es} (\href{https://orcid.org/0000-0001-7360-3249}{ORCID})

Authors:
\itemize{
  \item Manuel Oviedo de la Fuente \email{manuel.oviedo@udc.es} (\href{https://orcid.org/0000-0001-7360-3249}{ORCID})
}

}
\keyword{internal}
